refactor: swagger 패키지 분리
