+++ src/main/kotlin/com/yapp/lettie/api/timecapsule/controller/TimeCapsuleApiController.kt
@@ -5,7 +5,7 @@ import com.yapp.lettie.api.timecapsule.controller.request.CreateTimeCapsuleReque
 import com.yapp.lettie.api.timecapsule.controller.response.CreateTimeCapsuleResponse
 import com.yapp.lettie.api.timecapsule.controller.response.ToggleTimeCapsuleLikeResponse
 import com.yapp.lettie.api.timecapsule.service.TimeCapsuleService
-import com.yapp.lettie.api.timecapsule.swagger.TimeCapsuleSwagger
+import com.yapp.lettie.api.timecapsule.controller.swagger.TimeCapsuleSwagger
 import com.yapp.lettie.common.dto.ApiResponse
 import com.yapp.lettie.common.dto.UserInfoDto
 import org.springframework.http.ResponseEntity
